requirements.txt
Django==3.2.7
google-api-python-client
google-cloud-pubsub==2.9.0
python-dotenv==0.19.2
ortools==9.2.9972
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.8.6
numpy==1.21.5
scipy
google-cloud-storage==1.43.0